Working tFHE integration with JNI

+++ kvm-root/src/main/kotlin/kvm/instruction/KVEInstruction.kt
@@ -1,6 +1,6 @@
 package kvm.instruction
 
 sealed class KVEInstruction {
-    data class GreaterThan(val field: String, val value: Int) : KVEInstruction()
+    data class GreaterThan(val field: String, val value: Long) : KVEInstruction()
     data class Equals(val field: String, val value: String) : KVEInstruction()
 }
